+++ packages/react/.storybook/preview.js
@@ -9,7 +9,7 @@ export const parameters = {
   },
   options: {
     storySort: {
-      order: ['Views', 'Facets', 'Blocks', 'Media', '*'],
+      order: ['Introduction', 'Views', 'Facets', 'Blocks', 'Media', '*'],
     },
   },
   viewMode: 'docs',