+++ packages/react/src/blocks/Card/Card.test-stories.tsx
@@ -1,7 +1,7 @@
 import { Parameters } from '@storybook/addons'
 
 import { Escape } from '../../lib'
-import { CardProps } from '../../props/card-properties'
+import { CardProps } from '../../props'
 import { Paragraph } from '../Paragraph/Paragraph'
 import { Card } from './Card'
 
@@ -18,7 +18,10 @@ export const CardJsonTest: typeof CardTemplate & {
 } = CardTemplate.bind({})
 
 CardJsonTest.args = {
-  card: [{ paragraph: 'eafa9885-7c39-4255-8289-06cb0f5f509f' }],
+  card: {
+    title: '',
+    body: [{ paragraph: 'eafa9885-7c39-4255-8289-06cb0f5f509f' }],
+  },
 }
 CardJsonTest.parameters = { chromatic: { disableSnapshot: true } }
 
@@ -28,9 +31,12 @@ export const CardJsxTest: typeof CardTemplate & {
 } = CardTemplate.bind({})
 
 CardJsxTest.args = {
-  card: [
-    <Paragraph key="t1" paragraph="7dc865ac-d474-408f-a522-9189ce1ccbe5" />,
-  ],
+  card: {
+    title: '',
+    body: [
+      <Paragraph key="t1" paragraph="7dc865ac-d474-408f-a522-9189ce1ccbe5" />,
+    ],
+  },
 }
 CardJsxTest.parameters = { chromatic: { disableSnapshot: true } }
 
@@ -40,10 +46,13 @@ export const CardEscapeTest: typeof CardTemplate & {
 } = CardTemplate.bind({})
 
 CardEscapeTest.args = {
-  card: [
-    <Escape key="t1" contentMeetsAccessibilityAndDesignStandards>
-      <span>dea8bcd5-e9d0-488e-9974-eae698606b93</span>
-    </Escape>,
-  ],
+  card: {
+    title: '',
+    body: [
+      <Escape key="t1" contentMeetsAccessibilityAndDesignStandards>
+        <span>dea8bcd5-e9d0-488e-9974-eae698606b93</span>
+      </Escape>,
+    ],
+  },
 }
 CardEscapeTest.parameters = { chromatic: { disableSnapshot: true } }