+++ packages/react/src/blocks/Card/Card.tsx
@@ -4,9 +4,11 @@ import { mergeClasses as cx, makeStyles } from '@fluentui/react-components'
 // todo: fix this import when it stabilizes
 import { Card as FluentCard } from '@fluentui/react-components/unstable'
 
+import { Overflow } from '../../inputs'
 import {
   Sequence,
   invalidCardContentItem,
+  key,
   rem,
   renderIfEscape,
   sx,
@@ -16,9 +18,10 @@ import {
 import { CardContentItemEntity, CardProps } from '../../props'
 import { renderIfDescriptionList } from '../DescriptionList/DescriptionList'
 import { renderIfFigure } from '../Figure/Figure'
-import { renderIfHeading } from '../Heading/Heading'
+import { Heading, renderIfHeading } from '../Heading/Heading'
 import { renderIfParagraph } from '../Paragraph/Paragraph'
 import { renderIfShortInputs } from '../ShortInputs/ShortInputs'
+import { renderIfTable } from '../Table/Table'
 import { renderIfTabs } from '../Tabs/Tabs'
 import { WidgetPropsOrElement, renderIfWidget } from './exemplars/Widget'
 
@@ -29,14 +32,18 @@ const CardContentItem = (o: CardContentItemEntity) =>
   renderIfTabs(o) ||
   renderIfShortInputs(o) ||
   renderIfDescriptionList(o) ||
+  renderIfTable(o) ||
   renderIfEscape(o) ||
   invalidCardContentItem(o)
 
 const useCardStyles = makeStyles({
   root: {
     boxSizing: 'border-box',
-    ...sx.padding(rem(12)),
     ...sx.border('1px', 'solid', 'transparent'),
+    '& > .fuib-CardContentItem:not(.fui-CardPreview)': {
+      ...sx.margin(0),
+    },
+    paddingBlockEnd: rem(4),
   },
   hc: {
     ...sx.borderColor('var(--colorNeutralForeground1)'),
@@ -50,12 +57,25 @@ const useCardStyles = makeStyles({
     marginBlockStart: rem(20),
     marginBlockEnd: rem(20),
   },
+  headingRow: {
+    display: 'flex',
+    paddingBlockStart: rem(8),
+  },
+  headingText: {
+    ...sx.flex(1, 1, 'auto'),
+    marginBlockStart: rem(4),
+  },
+  cardContentSpacing: {
+    paddingInlineStart: rem(16),
+    paddingInlineEnd: rem(16),
+  },
 })
 
 export const Card = ({ card, contextualVariant = 'block' }: CardProps) => {
   const commonStyles = useCommonStyles()
   const cardStyles = useCardStyles()
   const { themeName } = useFluentBlocksContext()
+  const id = key(card)
   return (
     <FluentCard
       className={cx(
@@ -67,11 +87,39 @@ export const Card = ({ card, contextualVariant = 'block' }: CardProps) => {
         contextualVariant === 'block' && cardStyles.blockCard,
         contextualVariant === 'layout' && cardStyles.layoutItemCard
       )}
-      tabIndex={0}
+      aria-labelledby={id}
     >
-      {Sequence<CardContentItemEntity>(card, CardContentItem, {
-        contextualVariant: 'card',
-      })}
+      <div
+        role="none"
+        className={`${cx(
+          cardStyles.cardContentSpacing,
+          cardStyles.headingRow
+        )} fuib-CardContentItem`}
+      >
+        <div
+          role="none"
+          className={cx(
+            cardStyles.headingText,
+            card.titleVisuallyHidden && commonStyles.visuallyHidden
+          )}
+        >
+          <Heading
+            paragraph={card.title}
+            level={3}
+            contextualVariant="card"
+            contextualId={id}
+          />
+        </div>
+        {card.actions && <Overflow overflow={card.actions} />}
+      </div>
+      <div
+        role="none"
+        className={`${cardStyles.cardContentSpacing} fuib-CardContentItem`}
+      >
+        {Sequence<CardContentItemEntity>(card.body, CardContentItem, {
+          contextualVariant: 'card',
+        })}
+      </div>
     </FluentCard>
   )
 }