+++ packages/react/src/inlines/Icon/Icon.tsx
@@ -1,8 +1,9 @@
-import { ReactElement } from 'react'
 import get from 'lodash/get'
-import { makeStyles } from '@fluentui/react-components'
-import { IconVariant, IconSize, IconProps } from '@fluent-blocks/schemas'
+import { ReactElement } from 'react'
+
 import basicIcons from '@fluent-blocks/basic-icons'
+import { IconProps, IconSize, IconVariant } from '@fluent-blocks/schemas'
+import { makeStyles } from '@fluentui/react-components'
 
 import { useFluentBlocksContext } from '../../lib'
 
@@ -32,7 +33,7 @@ function spriteHref(
   }
 }
 
-const iconToTextRatio = 1.16
+const iconToTextRatio = 1.25
 
 const useIconStyles = makeStyles({
   root: {
@@ -48,7 +49,7 @@ export const Icon = (props: IconProps) => {
   const iconStyles = useIconStyles()
   const { iconSpriteUrl } = useFluentBlocksContext()
   return (
-    <svg className={iconStyles.root} data-chromatic="ignore">
+    <svg className={`${iconStyles.root} fuib-Icon`} data-chromatic="ignore">
       <use href={spriteHref(icon, size, variant, iconSpriteUrl)} />
     </svg>
   )