+++ packages/react/src/lib/theme.ts
@@ -1,18 +1,64 @@
+import isObject from 'lodash/isObject'
+import isString from 'lodash/isString'
+
 import {
-  teamsLightTheme,
+  PaletteConfig,
+  hexColorsFromPalette,
+  hex_to_LCH,
+} from '@fluent-blocks/colors'
+import {
+  AccentScheme as NaturalAccentScheme,
+  ThemeName as NaturalThemeName,
+  Palette,
+} from '@fluent-blocks/schemas'
+import {
+  BrandVariants,
+  createDarkTheme,
+  createLightTheme,
   teamsDarkTheme,
   teamsHighContrastTheme,
-  webLightTheme,
+  teamsLightTheme,
   webDarkTheme,
   webHighContrastTheme,
+  webLightTheme,
 } from '@fluentui/react-components'
-import {
-  ThemeName as NaturalThemeName,
-  AccentScheme as NaturalAccentScheme,
-} from '@fluent-blocks/schemas'
 
 export type ThemeName = NaturalThemeName
 export type AccentScheme = NaturalAccentScheme
+
+function isPalette(o: any): o is Palette {
+  return isObject(o) && 'keyColor' in o
+}
+
+export function getFullLCHPalette({ keyColor, ...props }: Palette) {
+  return {
+    keyColor: isString(keyColor) ? hex_to_LCH(keyColor) : keyColor,
+    darkCp: 2 / 3,
+    lightCp: 1 / 3,
+    hueTorsion: 0,
+    ...props,
+  }
+}
+
+export const defaultPaletteConfig: PaletteConfig = {
+  // The nShades and range values are based on a brand color audit
+  nShades: 16,
+  range: [1.42, 83.57],
+  linearity: 0.77,
+}
+
+function getBrandTokensFromPalette(palette: Palette) {
+  const hexColors = hexColorsFromPalette(
+    getFullLCHPalette(palette),
+    defaultPaletteConfig.nShades,
+    defaultPaletteConfig.range,
+    defaultPaletteConfig.linearity
+  )
+  return hexColors.reduce((acc: Record<string, string>, hexColor, h) => {
+    acc[`${(h + 1) * 10}`] = hexColor
+    return acc
+  }, {}) as BrandVariants
+}
 
 export const getTheme = (
   themeName?: ThemeName,
@@ -20,29 +66,42 @@ export const getTheme = (
 ) => {
   const resolvedAccentScheme = accentScheme || 'web'
   const resolvedThemeName = themeName || 'light'
-  switch (resolvedAccentScheme) {
-    case 'teams':
-      return (() => {
-        switch (resolvedThemeName) {
-          case 'highContrast':
-            return teamsHighContrastTheme
-          case 'dark':
-            return teamsDarkTheme
-          default:
-            return teamsLightTheme
-        }
-      })()
-    default:
-      return (() => {
-        switch (resolvedThemeName) {
-          case 'highContrast':
-            return webHighContrastTheme
-          case 'dark':
-            return webDarkTheme
-          default:
-            return webLightTheme
-        }
-      })()
+  if (isPalette(resolvedAccentScheme)) {
+    if (resolvedThemeName === 'highContrast') {
+      return webHighContrastTheme
+    } else {
+      const brandTokens = getBrandTokensFromPalette(resolvedAccentScheme)
+      if (resolvedThemeName === 'light') {
+        return createLightTheme(brandTokens)
+      } else {
+        return createDarkTheme(brandTokens)
+      }
+    }
+  } else {
+    switch (resolvedAccentScheme) {
+      case 'teams':
+        return (() => {
+          switch (resolvedThemeName) {
+            case 'highContrast':
+              return teamsHighContrastTheme
+            case 'dark':
+              return teamsDarkTheme
+            default:
+              return teamsLightTheme
+          }
+        })()
+      default:
+        return (() => {
+          switch (resolvedThemeName) {
+            case 'highContrast':
+              return webHighContrastTheme
+            case 'dark':
+              return webDarkTheme
+            default:
+              return webLightTheme
+          }
+        })()
+    }
   }
 }