+++ packages/react/src/media/Chart/Chart.tsx
@@ -1,16 +1,17 @@
-import { ReactElement } from 'react'
 import { Chart as ChartJS } from 'chart.js'
 import set from 'lodash/set'
+import { ReactElement } from 'react'
+
 import { ChartProps } from '@fluent-blocks/schemas'
+import { makeStyles } from '@fluentui/react-components'
 
 import { invalidChart } from '../../lib'
-
-import { PieChart } from './variants/PieChart'
-import { VerticalBarChart } from './variants/VerticalBarChart'
+import { BubbleChart } from './variants/BubbleChart'
 import { HorizontalBarChart } from './variants/HorizontalBarChart'
 import { LineChart } from './variants/LineChart'
+import { PieChart } from './variants/PieChart'
 import { StackedLineChart } from './variants/StackedLineChart'
-import { BubbleChart } from './variants/BubbleChart'
+import { VerticalBarChart } from './variants/VerticalBarChart'
 
 set(ChartJS, 'defaults.global.legend.display', false)
 set(
@@ -19,7 +20,13 @@ set(
   `Segoe UI, system-ui, sans-serif`
 )
 
-export function Chart(props: ChartProps) {
+const useChartStyles = makeStyles({
+  root: {
+    '& canvas': { width: '100% !important' },
+  },
+})
+
+function ChartInner(props: ChartProps) {
   const { chart, label } = props
   switch (chart.type) {
     case 'pie':
@@ -45,6 +52,11 @@ export function Chart(props: ChartProps) {
     default:
       return invalidChart(props)
   }
+}
+
+export function Chart(props: ChartProps) {
+  const chartStyles = useChartStyles()
+  return <div className={chartStyles.root}>{ChartInner(props)}</div>
 }
 
 export type ChartElement = ReactElement<ChartProps, typeof Chart>