+++ packages/react/src/surfaces/Main/Main.tsx
@@ -1,8 +1,12 @@
-import { makeStyles, mergeClasses as cx } from '@fluentui/react-components'
-
-import { rem, useCommonStyles } from '../../lib'
+import { MainProps as NaturalMainProps } from '@fluent-blocks/schemas'
+import { mergeClasses as cx, makeStyles } from '@fluentui/react-components'
 
 import { Section, SectionContentProps } from '../../blocks'
+import { rem, useCommonStyles } from '../../lib'
+
+export interface MainProps
+  extends Pick<NaturalMainProps, 'variant'>,
+    SectionContentProps {}
 
 const useMainSectionStyles = makeStyles({
   root: {
@@ -14,9 +18,13 @@ const useMainSectionStyles = makeStyles({
     paddingInlineStart: '1rem',
     paddingInlineEnd: '1rem',
   },
+  'root--flush': {
+    paddingInlineStart: 0,
+    paddingInlineEnd: 0,
+  },
 })
 
-export const Main = (props: SectionContentProps) => {
+export const Main = (props: MainProps) => {
   const mainStyles = useMainSectionStyles()
   const commonStyles = useCommonStyles()
   return (
@@ -24,7 +32,11 @@ export const Main = (props: SectionContentProps) => {
       {...props}
       as="main"
       level={1}
-      className={cx(commonStyles.baseSurface, mainStyles.root)}
+      className={cx(
+        commonStyles.baseSurface,
+        mainStyles.root,
+        props.variant === 'flush' && mainStyles['root--flush']
+      )}
     />
   )
 }