+++ packages/schemas/types/blocks/Card.d.ts
@@ -1,20 +1,27 @@
-import { HeadingProps } from './Heading'
-import { ParagraphProps } from './Paragraph'
-import { FigureProps } from './Figure'
-import { TabsProps } from './Tabs'
-import { ShortInputsProps } from './ShortInputs'
+import { InlineSequenceOrString } from '../inlines'
+import { MenuActionSequence } from '../lib/menu'
 import { DescriptionListProps } from './DescriptionList'
+import { FigureProps } from './Figure'
+import { ParagraphProps } from './Paragraph'
+import { ShortInputsProps } from './ShortInputs'
+import { TableProps } from './Table'
+import { TabsProps } from './Tabs'
 
 export type CardContentItemEntity =
-  | HeadingProps
   | ParagraphProps
   | FigureProps
   | TabsProps
   | ShortInputsProps
   | DescriptionListProps
+  | TableProps
 
 export type CardContentItemSequence = CardContentItemEntity[]
 
 export interface CardProps {
-  card: CardContentItemSequence
+  card: {
+    title: InlineSequenceOrString
+    titleVisuallyHidden?: boolean
+    actions?: MenuActionSequence
+    body: CardContentItemSequence
+  }
 }