+++ packages/schemas/types/exemplars/Widget.d.ts
@@ -1,11 +1,11 @@
 import { TabsProps } from '../blocks'
-import { ButtonProps } from '../inputs'
 import { InlineSequenceOrString } from '../inlines'
+import { ButtonProps } from '../inputs'
 
 // Widget is an exemplar of Card.
 
 interface Widget extends Omit<TabsProps, 'tabVariant' | 'tabListVariant'> {
-  title?: InlineSequenceOrString
+  title: InlineSequenceOrString
   abstract?: InlineSequenceOrString
   footerAction?: Omit<ButtonProps, 'type' | 'variant' | 'iconOnly'>
 }