+++ packages/schemas/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @fluent-blocks/schemas
 
+## 9.1.0-next.3
+
+### Patch Changes
+
+- [#97](https://github.com/OfficeDev/fluent-blocks/pull/97) [`c7d1e68`](https://github.com/OfficeDev/fluent-blocks/commit/c7d1e68f2e0b12a8ad68c7d5a26697c5dbba1e77) Thanks [@thure](https://github.com/thure)! - Update logo.
+
+- Updated dependencies [[`6d8cae3`](https://github.com/OfficeDev/fluent-blocks/commit/6d8cae381bcc38274a39817a3264a1df020f8dea)]:
+  - @fluent-blocks/colors@9.1.0-next.1
+
+## 9.1.0-next.2
+
+### Minor Changes
+
+- [#95](https://github.com/OfficeDev/fluent-blocks/pull/95) [`b2690c4`](https://github.com/OfficeDev/fluent-blocks/commit/b2690c44e9bc8306544f759f702c24b92da3235f) Thanks [@thure](https://github.com/thure)! - Adjust card content constituents to handle 280px viewport widths. Add 'flush' variant to Main.
+
+## 9.1.0-next.1
+
+### Minor Changes
+
+- [`1502a39`](https://github.com/OfficeDev/fluent-blocks/commit/1502a39e6e225886b377b381349645e5f7763838) Thanks [@thure](https://github.com/thure)! - Automate prereleases.
+
 ## 9.0.0
 
 ### Major Changes